requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mysql-connector-python==8.1.0
bcrypt==4.0.1
python-dotenv==1.0.0
pydantic==2.5.0